Moved setDurationToPhase to ContinuousEffect so that it can be used for other things
